move (Before|After)Suite to suite_test
